requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
openai>=1.12.0
python-multipart==0.0.6
aiohttp==3.9.1
requests==2.32.5
sqlalchemy==2.0.25
langchain-google-genai
langchain-core
tenacity
google-api-core
google-generativeai
streamlit
pillow
